% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{solution}
\alias{solution}
\title{Solution}
\usage{
solution(data)
}
\arguments{
\item{data}{The data.frame must have seven variables in the following order: 'age groups (AG)', 'number of women by AG', 'location', 'number of live births from women in each AG', number of surviving childern from women in each AG', 'children sex', 'year'.}
}
\description{
Put description
}
