% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutionFilter.R
\name{solutionFilter}
\alias{solutionFilter}
\title{Solution With Filter}
\usage{
solutionFilter(data, locale = "none", time = "none", gender = "none")
}
\arguments{
\item{data}{The data.frame must have seven variables in the following order: 'age groups (AG)', 'number of women by AG', 'location', 'number of live births from women in each AG', number of surviving childern from women in each AG', 'children sex', 'year'.}

\item{locale}{the location you want to filter. The standard is none}

\item{time}{the year you want to filter. The standard is none}

\item{gender}{the sex you want to filter. The standard is none}
}
\description{
Testando documentação
}
